+++ R/vsm.getSusceptibility.R
@@ -4,6 +4,7 @@
 #'
 #' @param obj VSMdata object
 #' @param sweepDirection +1 or -1 for the direction to use for fit (-1 default)
+#' @param singleLoop if \code{TRUE}, will evaluate only one loop regardless of temperature
 #' @return VSM data frame with sweepData column
 #' @examples
 #' filename = vsm.getSampleFiles()
@@ -11,7 +12,7 @@
 #' d1 = vsm.getLoop(d,direction=-1)
 #' vsm.getSusceptibility(d1, direction=-1)
 #' @export
-vsm.getSusceptibility <- function(obj, sweepDirection = -1) {
+vsm.getSusceptibility <- function(obj, sweepDirection = -1, singleLoop=FALSE) {
 
   d = data.frame(
     H = obj@H,
@@ -20,13 +21,20 @@ vsm.getSusceptibility <- function(obj, sweepDirection = -1) {
     dir = obj@dir
   )
 
-  sapply(split(d, d$Temp), function(x) {
-    x1 = subset(x, H>0.6*max(x$H) & H<0.95*max(x$H) & dir == sweepDirection)
-    if(nrow(x1)>10){
-      lm(x1$M ~ x1$H) -> fit
-      signif(summary(fit)$coeff[c(2,4)],3)
-    } else {
-      c(NA,NA)
-    }
-  })
+  if (singleLoop) {
+    x1 = subset(d, H>0.6*max(d$H) & H<0.95*max(d$H))
+    lm(x1$M ~ x1$H) -> fit
+    r = signif(summary(fit)$coeff[c(2,4)],4)
+  } else {
+    sapply(split(d, d$Temp), function(x) {
+      x1 = subset(x, H>0.6*max(x$H) & H<0.95*max(x$H) & dir == sweepDirection)
+      if(nrow(x1)>10){
+        lm(x1$M ~ x1$H) -> fit
+        r = signif(summary(fit)$coeff[c(2,4)],4)
+      } else {
+        r = c(NA,NA)
+      }
+    })
+  }
+  r
 }