+++ tests/testthat/test-vsm-hyst-stats.R
@@ -4,7 +4,7 @@ test_that("check coercivity from hysteresis loop", {
   a = vsm.hystStatsLoop(d)
 
   expect_equal(a$Mrem, -39e-6, tolerance = 1e-3)
-  expect_equal(a$Hc, 413, tolerance = 1e-3)
+  expect_equal(a$Hc, 413, tolerance = 5e-3)
 })
 
 test_that("test correction of hysteresis loop data", {