+++ NEWS.md
@@ -1,5 +1,6 @@
 # quantumPPMS 0.3
 
+* vsm.getSusceptbility can work with one loop alone
 * introduce PPMSdata S4 class
 
 # quantumPPMS 0.2.1