+++ NEWS.md
@@ -1,5 +1,6 @@
 # quantumPPMS 0.3
 
+* vsm.help() provides help on how to transition from v0.2 to v0.3 functions
 * vsm.getSusceptbility can work with one loop alone
 * introduce PPMSdata S4 class